requirements.txt
Flask
requests
flask-cors
gunicorn==20.1.0
pymongo
python-dotenv

